Removed comments from Game class. Reformatted CharacterModel. Updated constructor for Character/Monster in own classes and in Main
