+++ src/main/java/com/monsters/model/Character.java
@@ -1,36 +1,9 @@
 package com.monsters.model;
 
-public class Character {
+public class Character extends CharacterModel {
 
-  private final String name;
-  private int health = 100;
-
-  private int keys;
-
-  public Character(String name) {
-    this.name = name;
-  }
-
-  public int attack(int damage){
-    // TODO add rng for random damage
-    System.out.println("Attack!");
-    return damage;
-  }
-
-  public void decreaseHealth(int damage) {
-    this.health -= damage;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public int getHealth() {
-    return health;
-  }
-
-  public void setHealth(int health) {
-    this.health = health;
+  public Character(String name, int health) {
+    super(name, health);
   }
 
   public int getKeys() {
@@ -39,5 +12,17 @@ public class Character {
 
   public void setKeys(int keys) {
     this.keys = keys;
+  }
+
+  @Override
+  public int attack(int damage) {
+    // TODO add rng for random damage
+    System.out.println("Attack!");
+    return damage;
+  }
+
+  @Override
+  public void speak() {
+    System.out.println("I will prevail");
   }
 }
