+++ src/main/java/com/monsters/model/Monster.java
@@ -1,9 +1,9 @@
 package com.monsters.model;
 
-public class Monster extends AbstractMonster{
+public class Monster extends CharacterModel {
 
-  public Monster(String name,String key) {
-    super(name);
+  public Monster(String name, int health, int key) {
+    super(name, health, key);
   }
 
   @Override
@@ -14,11 +14,8 @@ public class Monster extends AbstractMonster{
   }
 
   @Override
-  public void roar() {
+  public void speak() {
     System.out.println("Roar!");
 
   }
-
-
-
 }
